Swap incorrect test files
We managed to end up with `test/deep-map` testing `digitise`, and `test/digitise` testing `deep-map`. We therefore swap the two files around to correct this.
